Refactoring all instances of u64/usize to the new umem/imem model

+++ memflow/src/mem/cache/cached_memory_access.rs
@@ -157,7 +157,7 @@ impl<'a, T: PhysicalMemory, Q: CacheValidator> PhysicalMemory for CachedMemoryAc
                     let mut cached_page = cache.cached_page_mut(paddr, false);
                     if let PageValidity::Valid(buf) = &mut cached_page.validity {
                         // write-back into still valid cache pages
-                        let start = paddr - cached_page.address;
+                        let start = (paddr - cached_page.address) as usize;
                         buf[start..(start + data_chunk.len())].copy_from_slice(data_chunk.into());
                     }
